Moved IIndexDocument interface.
